requirements.txt
openai
pydantic
python-dotenv
dolphindb
jinja2